Added autoselect support between "docker compose" and "docker-compose"

+++ templates/ever-node/start_node.sh
@@ -1,4 +1,22 @@
 #!/bin/bash
+
+dc=""
+
+docker-compose version >/dev/null 2>/dev/null
+if [[ "$?" == "0" ]]; then
+    dc="docker-compose"
+fi
+
+docker compose version >/dev/null 2>/dev/null
+if [[ "$?" == "0" ]]; then
+    dc="docker compose"
+fi
+
+if [[ "$dc" == "" ]]; then
+    echo "Error: functioning docker compose or docker-compose not found"
+    exit 1
+fi
+
 cd deploy/ever-node
 rm -rf build/ever-node
 cd build && git clone --recursive {{EVERNODE_GITHUB_REPO}} ever-node
@@ -12,14 +30,14 @@ cd ../../
 echo "==============================================================================="
 echo "INFO: starting node on {{HOSTNAME}}..."
 
-docker-compose up --build -d
+$dc up --build -d
 docker exec --tty ever-node "/ever-node/scripts/generate_console_config.sh"
 
-docker-compose down -t 300
+$dc down -t 300
 sed -i 's/"client_enabled":.*/"client_enabled": true,/' configs/config.json
 sed -i 's/"service_enabled":.*/"service_enabled": true,/' configs/config.json
 sed -i 's/command: \["bash"\]/command: ["normal"]/' docker-compose.yml
-docker-compose up -d
+$dc up -d
 echo "INFO: starting node on {{HOSTNAME}}... DONE"
 echo "==============================================================================="
